pyproject.toml
[tool.poetry]
name = "takedown-test"
version = "0.1.0"
description = ""
authors = ["Abhinav Pandey <abhinav.pandey.1512@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
locust = "^2.24.1"
python-dotenv = "^1.0.1"
python-socketio = "^5.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
